hw/system/spatz_cluster/test/bootrom.S: Enable msip in bootrom
* In the current standalone tb, snitch cores moves out of the low-power
state (wfi) via debug requests. While this is fine for both standalone
and in-system behaviour, it is recommended that the sw stack supports
interrupt-based wake up. To do that, we enable the MSIP interrupt by
writing to its dedicated CSR. As an additional improvement, we modify
the name of the generic `exception` to `dummy_exception` to emphasize
its role as safe guardrail